formula and bug fixes

+++ Envirowear/src/components/Average.java
@@ -2,7 +2,7 @@ package components;
 
 public class Average {
 
-    public float getAverage(float senosrTemp1, float sensorTemp2){
+    public double getAverage(double senosrTemp1, double sensorTemp2){
 
         return (senosrTemp1+sensorTemp2)/2;
